Redis async improvements
