+++ Splitio-tests/Integration Tests/Impressions/RedisImpressionsCacheTests.cs
@@ -1,9 +1,9 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Newtonsoft.Json;
 using Splitio.Redis.Services.Cache.Classes;
 using Splitio.Redis.Services.Cache.Interfaces;
 using Splitio.Redis.Services.Domain;
 using Splitio.Telemetry.Domain;
+using Splitio_Tests.Resources;
 using System.Collections.Generic;
 
 namespace Splitio_Tests.Integration_Tests.Impressions
@@ -13,7 +13,7 @@ namespace Splitio_Tests.Integration_Tests.Impressions
     {
         private const string RedisPrefix = "test-mtks:";
 
-        private readonly IRedisAdapter _redisAdapter;
+        private readonly RedisAdapterForTests _redisAdapter;
         private readonly IImpressionsCache _impressionsCache;
 
         public RedisImpressionsCacheTests()
@@ -32,7 +32,7 @@ namespace Splitio_Tests.Integration_Tests.Impressions
             };
             var connectionPoolManager = new ConnectionPoolManager(config);
 
-            _redisAdapter = new RedisAdapter(config, connectionPoolManager);
+            _redisAdapter = new RedisAdapterForTests(config, connectionPoolManager);
             _impressionsCache = new RedisImpressionsCache(_redisAdapter, "ip", "version", "mm", RedisPrefix);
         }
