+++ Splitio.Redis/Services/Cache/Classes/RedisSegmentCache.cs
@@ -1,6 +1,5 @@
 ﻿using Splitio.Redis.Services.Cache.Interfaces;
 using Splitio.Services.Cache.Interfaces;
-using StackExchange.Redis;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -13,13 +12,18 @@ namespace Splitio.Redis.Services.Cache.Classes
         protected const string SegmentNameKeyPrefix = "segment.{segmentname}.";
         protected const string SegmentsKeyPrefix = "segments.";
 
-        public RedisSegmentCache(IRedisAdapter redisAdapter, string userPrefix = null) : base(redisAdapter, userPrefix) { }
+        private readonly IRedisAdapterConsumer _redisAdapterConsumer;
+
+        public RedisSegmentCache(IRedisAdapterConsumer redisAdapter, string userPrefix = null) : base(userPrefix)
+        {
+            _redisAdapterConsumer = redisAdapter;
+        }
 
         #region Consumer
         public List<string> GetSegmentNames()
         {
             var key = $"{RedisKeyPrefix}{SegmentsKeyPrefix}registered";
-            var result = _redisAdapter.SMembers(key);
+            var result = _redisAdapterConsumer.SMembers(key);
 
             return result.Select(x => (string)x).ToList();
         }
@@ -27,7 +31,7 @@ namespace Splitio.Redis.Services.Cache.Classes
         public List<string> GetSegmentKeys(string segmentName)
         {
             var key = $"{RedisKeyPrefix}{SegmentKeyPrefix}{segmentName}";
-            var keys = _redisAdapter.SMembers(key);
+            var keys = _redisAdapterConsumer.SMembers(key);
 
             if (keys == null)
                 return new List<string>();
@@ -39,13 +43,13 @@ namespace Splitio.Redis.Services.Cache.Classes
         {
             var redisKey = $"{RedisKeyPrefix}{SegmentKeyPrefix}{segmentName}";
 
-            return _redisAdapter.SIsMember(redisKey, key);
+            return _redisAdapterConsumer.SIsMember(redisKey, key);
         }
 
         public long GetChangeNumber(string segmentName)
         {
             var key = RedisKeyPrefix + SegmentNameKeyPrefix.Replace("{segmentname}", segmentName) + "till";
-            var changeNumberString = _redisAdapter.Get(key);
+            var changeNumberString = _redisAdapterConsumer.Get(key);
             var result = long.TryParse(changeNumberString, out long changeNumberParsed);
 
             return result ? changeNumberParsed : -1;
@@ -64,7 +68,7 @@ namespace Splitio.Redis.Services.Cache.Classes
         public async Task<long> GetChangeNumberAsync(string segmentName)
         {
             var key = RedisKeyPrefix + SegmentNameKeyPrefix.Replace("{segmentname}", segmentName) + "till";
-            var changeNumberString = await _redisAdapter.GetAsync(key);
+            var changeNumberString = await _redisAdapterConsumer.GetAsync(key);
             var result = long.TryParse(changeNumberString, out long changeNumberParsed);
 
             return result ? changeNumberParsed : -1;
@@ -73,7 +77,7 @@ namespace Splitio.Redis.Services.Cache.Classes
         public async Task<List<string>> GetSegmentKeysAsync(string segmentName)
         {
             var key = $"{RedisKeyPrefix}{SegmentKeyPrefix}{segmentName}";
-            var keys = await _redisAdapter.SMembersAsync(key);
+            var keys = await _redisAdapterConsumer.SMembersAsync(key);
 
             if (keys == null) return new List<string>();
 
@@ -85,7 +89,7 @@ namespace Splitio.Redis.Services.Cache.Classes
         public async Task<List<string>> GetSegmentNamesAsync()
         {
             var key = $"{RedisKeyPrefix}{SegmentsKeyPrefix}registered";
-            var result = await _redisAdapter.SMembersAsync(key);
+            var result = await _redisAdapterConsumer.SMembersAsync(key);
 
             return result
                 .Select(x => (string)x)
@@ -96,7 +100,7 @@ namespace Splitio.Redis.Services.Cache.Classes
         {
             var redisKey = $"{RedisKeyPrefix}{SegmentKeyPrefix}{segmentName}";
 
-            return await _redisAdapter.SIsMemberAsync(redisKey, key);
+            return await _redisAdapterConsumer.SIsMemberAsync(redisKey, key);
         }
 
         public Task<int> SegmentKeysCountAsync()
@@ -107,21 +111,6 @@ namespace Splitio.Redis.Services.Cache.Classes
         public Task<int> SegmentsCountAsync()
         {
             return Task.FromResult(0);
-        }
-        #endregion
-
-        #region Only for tests
-        public long RegisterSegment(string segmentName)
-        {
-            return RegisterSegments(new List<string>() { segmentName });
-        }
-
-        public long RegisterSegments(List<string> segmentNames)
-        {
-            var key = $"{RedisKeyPrefix}{SegmentsKeyPrefix}registered";
-            var segments = segmentNames.Select(x => (RedisValue)x).ToArray();
-
-            return _redisAdapter.SAdd(key, segments);
         }
         #endregion
     }
